Refactoring the class and package for viewmodel
