+++ test/controllers/AddATrusteeControllerSpec.scala
@@ -20,12 +20,13 @@ import base.SpecBase
 import forms.AddATrusteeFormProvider
 import models.{AddATrustee, FullName, IndividualOrBusiness, NormalMode, UserAnswers}
 import navigation.{FakeNavigator, Navigator}
-import pages.{AddATrusteePage, TrusteeIndividualOrBusinessPage, TrusteesNamePage}
+import pages.{AddATrusteePage, TrusteeComplete, TrusteeIndividualOrBusinessPage, TrusteesNamePage}
 import play.api.inject.bind
 import play.api.mvc.Call
 import play.api.test.FakeRequest
 import play.api.test.Helpers._
 import viewmodels.AddRow
+import viewmodels.Tag.Completed
 import views.html.AddATrusteeView
 
 class AddATrusteeControllerSpec extends SpecBase {
@@ -45,8 +46,10 @@ class AddATrusteeControllerSpec extends SpecBase {
   val userAnswersWithTrusteesComplete = UserAnswers(userAnswersId)
     .set(TrusteesNamePage(0), FullName("First 0", None, "Last 0")).success.value
     .set(TrusteeIndividualOrBusinessPage(0), IndividualOrBusiness.Individual).success.value
+    .set(TrusteeComplete(0), Completed).success.value
     .set(TrusteesNamePage(1), FullName("First 1", None, "Last 1")).success.value
     .set(TrusteeIndividualOrBusinessPage(1), IndividualOrBusiness.Individual).success.value
+    .set(TrusteeComplete(1), Completed).success.value
 
   "AddATrustee Controller" must {
